// src/api/generated/models/userResource.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * ANZIANO-API
 * OpenAPI spec version: 1.0.0
 */
import type { UserResourceUserType } from './userResourceUserType';
import type { UserResourceUserTypeLabel } from './userResourceUserTypeLabel';
import type { UserResourceFullPathProfilePicture } from './userResourceFullPathProfilePicture';
import type { UserResourceEmailVerifiedAt } from './userResourceEmailVerifiedAt';

export interface UserResource {
  /** ID */
  id: string;
  name: string;
  email: string;
  user_type: UserResourceUserType;
  user_type_label: UserResourceUserTypeLabel;
  role_id: string;
  full_path_profile_picture?: UserResourceFullPathProfilePicture;
  email_verified_at?: UserResourceEmailVerifiedAt;
  created_at: string;
  updated_at: string;
}